requirements.txt
celery==5.2.7
django==4.1
djangorestframework==3.13.1
requests==2.28.1
beautifulsoup4==4.11.1
redis==4.3.4
kombu==5.2.4

